+++ src/main/java/com/jtyjy/finance/manager/dto/individual/IndividualImportErrorDTO.java
@@ -5,6 +5,7 @@ import com.alibaba.excel.annotation.format.DateTimeFormat;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -123,7 +124,7 @@ public class IndividualImportErrorDTO extends ImportErrorDTO{
 
     @ApiModelProperty(value = "年额度")
     @ExcelProperty(value = "*年额度")
-    private Integer annualQuota;
+    private BigDecimal annualQuota;
 
     @ApiModelProperty(value = "备注")
     @ExcelProperty(value = "备注")
